requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
aiosqlite>=0.19.0
pysnmp-lextudio>=6.0.0
pyyaml>=6.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
aiosmtplib>=3.0.0
python-multipart>=0.0.6
